Version: 0.2.1u (task 5.6 — Подключить mypy + cleanup — done; subtasks 5.6.1 + 5.6.2).
Mypy подключён с пермиссивным baseline (mypy.ini: disallow_untyped_defs=False,
strict_optional=False, ignore_missing_imports=True). Запуск: `mypy .`.
Dev-deps вынесены в requirements-dev.txt. На момент подключения 0 ошибок
(было 29 в 8 файлах перед cleanup).

5.6.1 BugFix: state.AdventureSession.end_ts. Тип был Optional[datetime], а на
запись/чтение всегда использовался Unix timestamp (float). _deser_datetime(float)
возвращал None → adventure_end_timestamp терялся при каждом load. Игрок терял
прогресс активного приключения после reload/restart. Тип переведён на
Optional[float], from_dict читает значение напрямую. start_adventure() сигнатура
обновлена. 2 новых теста (test_adventure_end_ts_is_float,
test_adventure_end_ts_none_when_missing).

5.6.2 Cleanup: gspread ValueInputOption enum вместо устаревшей строки
'USER_ENTERED' в StepsLogRepo.append() — снимает DeprecationWarning.

Системный cleanup типов в bonus.py / gym.py / inventory.py / drop.py /
equipment.py / web/main.py / characteristics.py / adventure.py: int *= float
исправления через явный int() cast или новую переменную, dict[str, Any]
аннотации для полиморфных dict, getattr() с явным типом.

Документация: CLAUDE.md обновлён (Common commands + Type hints section),
changelog 0.2.1u, TASKS.md 5.6 / 5.6.1 / 5.6.2 → done.

Тесты: 395 passed. Mypy: Success, no issues found in 49 source files.

+++ inventory.py
@@ -204,7 +204,7 @@ class Wear_Equipped_Items:
 
     def reduce_wear(self, steps: int) -> None:
         """Уменьшает износ предметов с дополнительным учётом neatness (legacy API)."""
-        reduced_steps = steps * (1 - (self._state.gym.neatness_in_using_things / 100))
+        reduced_steps = int(steps * (1 - (self._state.gym.neatness_in_using_things / 100)))
         self.decrease_durability(reduced_steps)
 
     def view_wear_reduce_change(self, item_name: str, initial_quality: float,
